refactor: Migrate slash commands to cogs architecture
