--- src/commands/__init__.py
@@ -1,1 +1,0 @@
-"""Discord command registration modules."""
